店铺下有员工，无法删除

+++ src/main/java/com/linkknown/crm/service/impl/ShopServiceImpl.java
@@ -1,9 +1,11 @@
 package com.linkknown.crm.service.impl;
 
+import com.linkknown.crm.bean.dos.Employee;
 import com.linkknown.crm.bean.dos.Shop;
 import com.linkknown.crm.common.aspect.exception.WebException;
 import com.linkknown.crm.common.enums.ResponseEnum;
 import com.linkknown.crm.common.enums.SelectedEnum;
+import com.linkknown.crm.mapper.EmployeeMapper;
 import com.linkknown.crm.mapper.ShopMapper;
 import com.linkknown.crm.service.IShopService;
 import org.springframework.stereotype.Service;
@@ -22,6 +24,9 @@ public class ShopServiceImpl implements IShopService {
 
     @Resource
     private ShopMapper shopMapper;
+
+    @Resource
+    private EmployeeMapper employeeMapper;
 
 
     /**
@@ -117,10 +122,21 @@ public class ShopServiceImpl implements IShopService {
      */
     @Override
     public void deleteShop(Shop shop) {
+        //查看是否正在被选择
         Shop shopDb = shopMapper.selectShopById(Long.valueOf(shop.getShopId()));
         if (SelectedEnum.yes.getCode().equals(shopDb.getSelected())){
             throw new WebException(ResponseEnum.shop_is_selected_can_not_delete);
         }
+
+        //查看店铺下是否还有员工
+        Employee employee = new Employee();
+        employee.setShopId(shop.getShopId());
+        List<Employee> employeeList = employeeMapper.selectEmployeeList(employee);
+        if (!CollectionUtils.isEmpty(employeeList)){
+            throw new WebException(ResponseEnum.shop_has_employee_can_not_delete);
+        }
+
+        //删除
         shopMapper.deleteShopById(Long.valueOf(shop.getShopId()));
     }
